// portal/proto/isuxportal/resources/leaderboard_pb.d.ts
// @generated by protoc-gen-es v2.0.0
// @generated from file isuxportal/resources/leaderboard.proto (package isuxportal.proto.resources, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv1";
import type { Message } from "@bufbuild/protobuf";
import type { Team } from "./team_pb";
import type { Timestamp } from "@bufbuild/protobuf/wkt";
import type { Contest } from "./contest_pb";

/**
 * Describes the file isuxportal/resources/leaderboard.proto.
 */
export declare const file_isuxportal_resources_leaderboard: GenFile;

/**
 * @generated from message isuxportal.proto.resources.LeaderboardItem
 */
export declare type LeaderboardItem = Message<"isuxportal.proto.resources.LeaderboardItem"> & {
  /**
   * @generated from field: isuxportal.proto.resources.LeaderboardItem.LeaderboardScore best_score = 2;
   */
  bestScore?: LeaderboardItem_LeaderboardScore;

  /**
   * @generated from field: isuxportal.proto.resources.LeaderboardItem.LeaderboardScore latest_score = 3;
   */
  latestScore?: LeaderboardItem_LeaderboardScore;

  /**
   * @generated from field: isuxportal.proto.resources.Team team = 16;
   */
  team?: Team;

  /**
   * @generated from field: isuxportal.proto.resources.LeaderboardItem.History score_history = 17;
   */
  scoreHistory?: LeaderboardItem_History;
};

/**
 * Describes the message isuxportal.proto.resources.LeaderboardItem.
 * Use `create(LeaderboardItemSchema)` to create a new message.
 */
export declare const LeaderboardItemSchema: GenMessage<LeaderboardItem>;

/**
 * @generated from message isuxportal.proto.resources.LeaderboardItem.LeaderboardScore
 */
export declare type LeaderboardItem_LeaderboardScore = Message<"isuxportal.proto.resources.LeaderboardItem.LeaderboardScore"> & {
  /**
   * @generated from field: int64 score = 1;
   */
  score: bigint;

  /**
   * @generated from field: google.protobuf.Timestamp started_at = 2;
   */
  startedAt?: Timestamp;

  /**
   * @generated from field: google.protobuf.Timestamp marked_at = 3;
   */
  markedAt?: Timestamp;
};

/**
 * Describes the message isuxportal.proto.resources.LeaderboardItem.LeaderboardScore.
 * Use `create(LeaderboardItem_LeaderboardScoreSchema)` to create a new message.
 */
export declare const LeaderboardItem_LeaderboardScoreSchema: GenMessage<LeaderboardItem_LeaderboardScore>;

/**
 * @generated from message isuxportal.proto.resources.LeaderboardItem.History
 */
export declare type LeaderboardItem_History = Message<"isuxportal.proto.resources.LeaderboardItem.History"> & {
  /**
   * @generated from field: repeated isuxportal.proto.resources.LeaderboardItem.LeaderboardScore scores = 1;
   */
  scores: LeaderboardItem_LeaderboardScore[];
};

/**
 * Describes the message isuxportal.proto.resources.LeaderboardItem.History.
 * Use `create(LeaderboardItem_HistorySchema)` to create a new message.
 */
export declare const LeaderboardItem_HistorySchema: GenMessage<LeaderboardItem_History>;

/**
 * @generated from message isuxportal.proto.resources.Leaderboard
 */
export declare type Leaderboard = Message<"isuxportal.proto.resources.Leaderboard"> & {
  /**
   * @generated from field: repeated isuxportal.proto.resources.LeaderboardItem teams = 1;
   */
  teams: LeaderboardItem[];

  /**
   * @generated from field: repeated isuxportal.proto.resources.LeaderboardItem hidden_teams = 7;
   */
  hiddenTeams: LeaderboardItem[];

  /**
   * @generated from field: repeated isuxportal.proto.resources.LeaderboardItem progresses = 4;
   */
  progresses: LeaderboardItem[];

  /**
   * @generated from field: google.protobuf.Timestamp generated_at = 6;
   */
  generatedAt?: Timestamp;

  /**
   * @generated from field: isuxportal.proto.resources.Contest contest = 5;
   */
  contest?: Contest;
};

/**
 * Describes the message isuxportal.proto.resources.Leaderboard.
 * Use `create(LeaderboardSchema)` to create a new message.
 */
export declare const LeaderboardSchema: GenMessage<Leaderboard>;